@@ -22,18 +22,18 @@
     "bufferutil": "^4.0.7"
   },
   "devDependencies": {
+    "@types/better-sqlite3": "^7.6.5",
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^20.10.6",
-    "eslint": "^8.50.0",
-    "typescript": "^5.3.3",
-    "prettier": "^3.0.3",
+    "@typescript-eslint/eslint-plugin": "^6.18.0",
+    "@typescript-eslint/parser": "^6.18.0",
+    "eslint": "^8.56.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^5.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
-    "@types/better-sqlite3": "^7.6.5"
+    "prettier": "^3.0.3",
+    "typescript": "^5.3.3"
   }
 }
