@@ -1,16 +1,46 @@
 // DEFINITIONS
+import { jsonc } from 'jsonc';
+import fs from 'fs';
+import path from 'path';
+
+// LOCALES DEFINITION //
+const locales: any = jsonc.parse(fs.readFileSync(path.join(__dirname, '../locales/locales.jsonc'), 'utf8'));
+// DATABASE DEFINITION //
+import database from '../index';
+// TICKET TABLE DEFINITION //
+const ticket = database.table('ticket');
+// SETTINGS DEFINITION //
+const settings: any = jsonc.parse(fs.readFileSync(path.join(__dirname, '../config/settings.jsonc'), 'utf8'));
+// VERSION //
+import { version } from '../package.json';
+
+// FUNCTIONS //
 import close from '../utils/close';
 import newTicket from '../utils/openTicket';
+import {getDatestamp, getTimestamp, hasNewUsername} from '../utils/etc'
+import { webhook } from '../utils/webhook';
 
-// LINKING UTILITY
-const lib = {};
+// TESTING FEATURE //
+const test = function(): boolean {
+	return true;
+};
 
-// lib.close = function(base, type, ticketNumber) {
-// 	close(base, type, ticketNumber);
-// };
-// lib.newTicket = function(base) {
-// 	newTicket(base);
-// };
+const lib = {
+	locales: locales,
+	db: database,
+	ticket: ticket,
+	settings: settings,
+	// FUNCTIONS //
+	close: close,
+	newTicket: newTicket,
+	hasNewUsername: hasNewUsername,
+	timestamp: getTimestamp,
+	datestamp: getDatestamp,
+	wbh: webhook,
+	version: version,
 
+	//TEST FUNCTION//
+	test: test
+};
 // EXPORT
 export default lib;
