@@ -1,17 +1,17 @@
-import { SlashCommandBuilder, EmbedBuilder } from 'discord.js';
+import { SlashCommandBuilder, EmbedBuilder, CommandInteraction } from 'discord.js';
+import lib from '../../bridge/bridge'
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('ping')
 		.setDescription('Preveri zakasnitev programa'),
-	async execute(interaction:any) {
-		const client = interaction.client;
+	async execute(interaction: CommandInteraction) {
 		const message = await interaction.deferReply({
 			ephemeral: true,
 			fetchReply: true,
 		});
 		const embed = new EmbedBuilder()
-			.setColor(await client.db.get('color.default'))
-			.setTitle(`${client.locales.commands.pingjs.embed.title} ${message.createdTimestamp - interaction.createdTimestamp}ms`);
+			.setColor(await lib.db.get('color.default'))
+			.setTitle(`${lib.locales.commands.pingjs.embed.title} ${message.createdTimestamp - interaction.createdTimestamp}ms`);
 		await interaction.editReply({
 			embeds: [embed],
 		});
