@@ -1,30 +1,30 @@
-import { SlashCommandBuilder, EmbedBuilder } from 'discord.js';
-import { Table } from 'quick.db';
+import { SlashCommandBuilder, EmbedBuilder, CommandInteraction, Client, TextChannel, DMChannel } from 'discord.js';
+import { QuickDB } from 'quick.db';
+import lib from '../../bridge/bridge'
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('cinactive')
 		.setDMPermission(false)
 		.setDescription('Odstrani oznako neaktivnosti'),
-	async execute(interaction:any) {
+	async execute(interaction: CommandInteraction) {
 		const client = interaction.client;
-		// const locales = client.locales.utils.closejs;
 		await interaction.deferReply({ ephemeral: true });
-		if (!await client.ticket.has(interaction.channelId)) {
+		if (!await lib.ticket.has(interaction.channelId)) {
 			interaction.editReply('Ta kanal ni aktiven ticket');
 			return;
 		}
-		const number = await client.ticket.get(interaction.channelId);
-		const database = await client.db.table(`tt_${number}`);
+		const number = await lib.ticket.get(interaction.channelId);
+		const database = lib.db.table(`tt_${number}`);
 		const data = await database.get('info');
-		const que_chc1 = await client.ticket.get('inaQueue');
+		const que_chc1 = await lib.ticket.get('inaQueue');
 		if (!que_chc1) {
-			await client.ticket.set('inaQueue', []);
+			await lib.ticket.set('inaQueue', []);
 			interaction.editReply('Ticket nima oznake neaktivnosti');
 		}
 		else {
-			const que_chc2 = await client.ticket.get('inaQueue');
+			const que_chc2 = await lib.ticket.get('inaQueue');
 			if (que_chc2.includes(number)) {
-				await client.ticket.pull('inaQueue', number);
+				await lib.ticket.pull('inaQueue', number);
 			}
 			else {
 				interaction.editReply('Ticket nima oznake neaktivnosti');
@@ -32,17 +32,18 @@ module.exports = {
 			}
 		}
 		const embed = new EmbedBuilder()
-			.setColor(await client.db.get('color.default'))
+			.setColor(await lib.db.get('color.default'))
 			.setTitle('Oznaka neaktivnosti odstranjena')
 			.setDescription('Ticket je bil označen kot aktiven to pomeni, \nda se ticket **ne** bo avtomatsko zaprl, če v 48ih urah ni sporočil.');
 
 		const emb = new EmbedBuilder()
-			.setColor(await client.db.get('color.default'))
+			.setColor(await lib.db.get('color.default'))
 			.setTitle('Oznaka neaktivnosti odstranjena')
 			.setDescription('Ticket je bil označen kot aktiven to pomeni, \nda se ticket **ne** bo avtomatsko zaprl, če v 48ih urah ni sporočil.');
 
 		const channels = data.dmChannel;
-		const channel = await client.channels.fetch(data.guildChannel);
+		const passedChannel = await client.channels.fetch(data.guildChannel);
+		const channel = passedChannel as TextChannel
 		sendEmbeds(client, channels, embed);
 		sendToServer(client, channel, emb);
 		setData(database);
@@ -50,10 +51,11 @@ module.exports = {
 	},
 };
 
-async function sendEmbeds(client:any, channels:any, embed:EmbedBuilder) {
+async function sendEmbeds(client: Client, channels: Array<any>, embed:EmbedBuilder) {
 	for (const id of channels) {
 		try {
-			const channel = await client.channels.fetch(id);
+			const passedChannel = await client.channels.fetch(id);
+			const channel = passedChannel as TextChannel | DMChannel
 			await channel.send({ embeds: [embed] });
 		}
 		catch (e) {
@@ -62,11 +64,11 @@ async function sendEmbeds(client:any, channels:any, embed:EmbedBuilder) {
 	}
 }
 
-async function sendToServer(client:any, channel:any, emb:EmbedBuilder) {
-	const wbh = await client.wbh(channel);
-	wbh.send({ embeds: [emb] });
+async function sendToServer(client: Client, channel: TextChannel, emb:EmbedBuilder) {
+	const wbh = await lib.wbh(channel);
+	wbh?.send({ embeds: [emb] });
 }
 
-async function setData(database:Table) {
+async function setData(database: QuickDB) {
 	database.delete('inaData');
 }
