@@ -1,48 +1,50 @@
-import { SlashCommandBuilder, EmbedBuilder } from 'discord.js';
-import { Table } from 'quick.db';
+import { SlashCommandBuilder, EmbedBuilder, CommandInteraction, Client, TextChannel, DMChannel } from 'discord.js';
+import { QuickDB } from 'quick.db';
+import lib from '../../bridge/bridge'
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('inactive')
 		.setDMPermission(false)
 		.setDescription('Oznaci kot inaktiven ticket'),
-	async execute(interaction:any) {
+	async execute(interaction: CommandInteraction) {
 		const client = interaction.client;
 		// const locales = client.locales.utils.closejs;
 		await interaction.deferReply({ ephemeral: true });
-		if (!await client.ticket.has(interaction.channelId)) {
+		if (!await lib.ticket.has(interaction.channelId)) {
 			interaction.editReply('Ta kanal ni aktiven ticket');
 			return;
 		}
-		const number = await client.ticket.get(interaction.channelId);
-		const database = await client.db.table(`tt_${number}`);
+		const number = await lib.ticket.get(interaction.channelId);
+		const database = lib.db.table(`tt_${number}`);
 		const data = await database.get('info');
-		const que_chc1 = await client.ticket.get('inaQueue');
+		const que_chc1 = await lib.ticket.get('inaQueue');
 		if (!que_chc1) {
-			await client.ticket.set('inaQueue', []);
-			await client.ticket.push('inaQueue', number);
+			await lib.ticket.set('inaQueue', []);
+			await lib.ticket.push('inaQueue', number);
 		}
 		else {
-			const que_chc2 = await client.ticket.get('inaQueue');
+			const que_chc2 = await lib.ticket.get('inaQueue');
 			if (que_chc2.includes(number)) {
 				interaction.editReply('Ticket je že označen kot neaktiven');
 				return;
 			}
 			else {
-				await client.ticket.push('inaQueue', number);
+				await lib.ticket.push('inaQueue', number);
 			}
 		}
 		const embed = new EmbedBuilder()
-			.setColor(await client.db.get('color.default'))
+			.setColor(await lib.db.get('color.default'))
 			.setTitle('Oznaka inaktivnosti')
 			.setDescription('Vaš ticket je bil označen kot neaktiven to pomeni, \nda mora biti v vašem ticketu poslano sporočilo vsakih \n48 ur drugače se bo ticket avtomatsko zaprl.\n24 ur pred zaprtjem boste opozorjeni.');
 
 		const emb = new EmbedBuilder()
-			.setColor(await client.db.get('color.default'))
+			.setColor(await lib.db.get('color.default'))
 			.setTitle('Oznaka inaktivnosti')
 			.setDescription('Ticket je bil označen kot neaktiven to pomeni, \nda se bo ticket avtomatsko zaprl, če v 48ih urah ni sporočil.\n24 ur pred zaprtjem boste opozorjeni.');
 
 		const channels = data.dmChannel;
-		const channel = await client.channels.fetch(data.guildChannel);
+		const passedChannel = await client.channels.fetch(data.guildChannel);
+		const channel = passedChannel as TextChannel
 		sendEmbeds(client, channels, embed);
 		sendToServer(client, channel, emb);
 		setData(database);
@@ -50,10 +52,11 @@ module.exports = {
 	},
 };
 
-async function sendEmbeds(client:any, channels:any, embed:EmbedBuilder) {
+async function sendEmbeds(client: Client, channels: Array<any>, embed:EmbedBuilder) {
 	for (const id of channels) {
 		try {
-			const channel = await client.channels.fetch(id);
+			const passedChannel = await client.channels.fetch(id);
+			const channel = passedChannel as TextChannel | DMChannel;
 			await channel.send({ embeds: [embed] });
 		}
 		catch (e) {
@@ -62,12 +65,14 @@ async function sendEmbeds(client:any, channels:any, embed:EmbedBuilder) {
 	}
 }
 
-async function sendToServer(client:any, channel:any, emb:EmbedBuilder) {
-	const wbh = await client.wbh(channel);
-	wbh.send({ embeds: [emb] });
+async function sendToServer(client: Client, channel: TextChannel | null, emb:EmbedBuilder) {
+	if (channel) {
+		const wbh = await lib.wbh(channel);
+		wbh?.send({ embeds: [emb] });
+	}
 }
 
-async function setData(database:Table) {
+async function setData(database:QuickDB) {
 	const now = new Date();
 	const currentSec = now.getSeconds();
 	const currentMs = now.getMilliseconds();
