@@ -1,4 +1,5 @@
-import { SlashCommandBuilder, EmbedBuilder, ButtonBuilder, ButtonStyle, ActionRowBuilder, User, DMChannel } from 'discord.js';
+import { SlashCommandBuilder, EmbedBuilder, ButtonBuilder, ButtonStyle, ActionRowBuilder, User, DMChannel, CommandInteraction, Client } from 'discord.js';
+import lib from '../../bridge/bridge'
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('add')
@@ -7,23 +8,25 @@ module.exports = {
 			option.setName('user')
 				.setDescription('Uporabnik, ki bo dodan v ticket')
 				.setRequired(true)),
-	async execute(interaction:any) {
+	async execute(interaction: CommandInteraction) {
 		const client = interaction.client;
-		const locales = client.locales.commands.adduserjs;
+		const locales = lib.locales.commands.adduserjs;
 		const user = interaction.options.getUser('user');
-		const checkOne = await client.ticket.has(interaction.channelId);
-		if (checkOne === false) {
-			interaction.reply(locales.notActiveChannel);
-			return;
+		if (user) {
+			const checkOne = await lib.ticket.has(interaction.channelId);
+			if (checkOne === false) {
+				interaction.reply(locales.notActiveChannel);
+				return;
+			}
+			userCheck(interaction, client, user);
 		}
-		userCheck(interaction, client, user);
 	},
 };
 
-async function userCheck(interaction:any, client:any, user:User) {
-	const locales = client.locales.commands.adduserjs;
-	const ticketDatabaseNumber = await client.ticket.get(interaction.channelId);
-	const ticketDatabase = await client.db.table(`tt_${ticketDatabaseNumber}`).get('info');
+async function userCheck(interaction: CommandInteraction, client: Client, user:User) {
+	const locales = lib.locales.commands.adduserjs;
+	const ticketDatabaseNumber = await lib.ticket.get(interaction.channelId);
+	const ticketDatabase = await lib.db.table(`tt_${ticketDatabaseNumber}`).get('info');
 	const status = await ticketDatabase.dmChannel.includes(user.id);
 
 	switch (status) {
@@ -36,13 +39,13 @@ async function userCheck(interaction:any, client:any, user:User) {
 	}
 }
 
-async function addUserToTicket(interaction:any,	user:User, num:number) {
-	const locales = interaction.client.locales.commands.adduserjs.addUserToTicket;
+async function addUserToTicket(interaction: CommandInteraction,	user:User, num:number) {
+	const locales = lib.locales.commands.adduserjs.addUserToTicket;
 	const DM = await user.createDM();
 	const hasTicket = await checkIfUserHasOpenTicket(interaction, DM);
-	const ticketCreator = await interaction.client.db.table(`tt_${num}`).get('info.creatorId');
+	const ticketCreator = await lib.db.table(`tt_${num}`).get('info.creatorId');
 	const creator = await interaction.client.users.fetch(ticketCreator);
-	const username = await interaction.client.hasNewUsername(creator, true, 'user');
+	const username = await lib.hasNewUsername(creator, true, 'user');
 	if (hasTicket === true) {
 		interaction.reply(locales.userHasTicket);
 		return;
@@ -71,12 +74,12 @@ async function addUserToTicket(interaction:any,	user:User, num:number) {
 		.replace('USERNAME', `<@${user.id}>`));
 }
 
-async function databaseSync(interaction:any, DM:DMChannel, num:number) {
-	await interaction.client.ticket.set(DM.id, num);
-	await interaction.client.db.table(`tt_${num}`).push('info.dmChannel', DM.id);
+async function databaseSync(interaction: CommandInteraction, DM:DMChannel, num:number) {
+	await lib.ticket.set(DM.id, num);
+	await lib.db.table(`tt_${num}`).push('info.dmChannel', DM.id);
 }
-async function checkIfUserHasOpenTicket(interaction:any, DM:DMChannel) {
-	const status = await interaction.client.ticket.has(DM.id);
+async function checkIfUserHasOpenTicket(interaction: CommandInteraction, DM:DMChannel) {
+	const status = await lib.ticket.has(DM.id);
 	if (status === true) return true;
 	if (status === false) return false;
 	return undefined;
