@@ -1,11 +1,11 @@
-import { SlashCommandBuilder } from 'discord.js';
+import { SlashCommandBuilder, CommandInteraction } from 'discord.js';
+import lib from '../../bridge/bridge'
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('close')
 		.setDescription('Zapri ticket')
 		.setDMPermission(false),
-	async execute(interaction:any) {
-		const lib = interaction.client.lib;
-		lib.close(interaction, 'cls');
+	async execute(interaction: CommandInteraction) {
+		lib.close(interaction, 'cls', null);
 	},
 };
