@@ -1,28 +1,29 @@
-import { Events, ChannelType, EmbedBuilder, TextChannel, CategoryChannel } from 'discord.js';
+import { Events, ChannelType, EmbedBuilder, TextChannel, CategoryChannel, Client, Guild, BaseGuild, Role, Snowflake } from 'discord.js';
+import lib from '../../bridge/bridge';
 module.exports = {
 	name: Events.ClientReady,
 	once: true,
-	async execute(client:any) {
-		const guild = await client.guilds.fetch(client.settings.guildId);
-		const status = await client.db.has(guild.id);
+	async execute(client: Client) {
+		const guild = await client.guilds.fetch(lib.settings.guildId);
+		const status = await lib.db.has(guild.id);
 		if (status === false) {
 			try {
 				const category = await guild.channels.create({
-					name: client.locales.events.channelsjs.channelCreation.category,
+					name: lib.locales.events.channelsjs.channelCreation.category,
 					type: ChannelType.GuildCategory,
 				});
 				const categoryClosed = await guild.channels.create({
-					name: client.locales.events.channelsjs.channelCreation.categoryClosed,
+					name: lib.locales.events.channelsjs.channelCreation.categoryClosed,
 					type: ChannelType.GuildCategory,
 				});
 
 				const log = await category.children.create({
-					name: client.locales.events.channelsjs.channelCreation.log,
+					name: lib.locales.events.channelsjs.channelCreation.log,
 					type: ChannelType.GuildText,
 				});
 
 				const archive = await category.children.create({
-					name: client.locales.events.channelsjs.channelCreation.archive,
+					name: lib.locales.events.channelsjs.channelCreation.archive,
 					type: ChannelType.GuildText,
 				});
 
@@ -31,7 +32,7 @@ module.exports = {
 				await log.permissionOverwrites.create(guild.roles.everyone, { ViewChannel: false });
 				await archive.permissionOverwrites.create(guild.roles.everyone, { ViewChannel: false });
 
-				permissionSet(category, log, archive, categoryClosed, client, guild);
+				permissionSet(category, log, archive, categoryClosed, guild);
 			}
 			catch (e) {
 				console.error(e);
@@ -41,30 +42,30 @@ module.exports = {
 	},
 };
 
-async function permissionSet(category:any, log:TextChannel, archive:TextChannel, categoryClosed:CategoryChannel, client:any, guild:any) {
-	const roles = client.settings.allowedRoles;
-
-	for (const role of roles) {
-		const x = await guild.roles.cache.get(role);
-		await category.permissionOverwrites.create(x, { ViewChannel: true });
-		await log.permissionOverwrites.create(x, { ViewChannel: true });
-		await archive.permissionOverwrites.create(x, { ViewChannel: true });
-		await categoryClosed.permissionOverwrites.create(x, { ViewChannel: true });
+async function permissionSet(category: CategoryChannel, log:TextChannel, archive:TextChannel, categoryClosed:CategoryChannel, guild: Guild) {
+	const roles = lib.settings.allowedRoles;
+	for (const roleLoop of roles) {
+		const x = guild.roles.cache.get(roleLoop);
+		const role = x as Role;
+		await category.permissionOverwrites.create(role, { ViewChannel: true });
+		await log.permissionOverwrites.create(role, { ViewChannel: true });
+		await archive.permissionOverwrites.create(role, { ViewChannel: true });
+		await categoryClosed.permissionOverwrites.create(role, { ViewChannel: true });
 	}
-	const wbh = await client.wbh(log);
-	const wbh2 = await client.wbh(archive);
+	const wbh = await lib.wbh(log);
+	const wbh2 = await lib.wbh(archive);
 
-	await client.db.set(guild.id, { 'logChannel': log.id, 'transcriptChannel': archive.id, 'categoryId': category.id, 'closeCategoryId': categoryClosed.id });
+	await lib.db.set(guild.id, { 'logChannel': log.id, 'transcriptChannel': archive.id, 'categoryId': category.id, 'closeCategoryId': categoryClosed.id });
 
 	const embed = new EmbedBuilder()
-		.setColor(await client.db.get('color.default'))
-		.setTitle(client.locales.events.channelsjs.logEmbed.title)
+		.setColor(await lib.db.get('color.default'))
+		.setTitle(lib.locales.events.channelsjs.logEmbed.title)
 		.setTimestamp();
 	const embed2 = new EmbedBuilder()
-		.setColor(await client.db.get('color.default'))
-		.setTitle(client.locales.events.channelsjs.archiveEmbed.title)
+		.setColor(await lib.db.get('color.default'))
+		.setTitle(lib.locales.events.channelsjs.archiveEmbed.title)
 		.setTimestamp();
 
-	wbh.send({ embeds: [embed] });
-	wbh2.send({ embeds: [embed2] });
+	wbh?.send({ embeds: [embed] });
+	wbh2?.send({ embeds: [embed2] });
 }
