@@ -1,13 +1,20 @@
-import { Events, EmbedBuilder, Message } from 'discord.js';
-import { Table } from 'quick.db';
+import { Events, EmbedBuilder, Message, Client, DMChannel, TextChannel } from 'discord.js';
+import { QuickDB } from 'quick.db';
+import lib from '../../bridge/bridge';
 module.exports = {
 	name: Events.MessageUpdate,
-	async execute(newMessage: any) {
+	async execute(oldMessage: Message, newMessage: Message) {
 		const client = newMessage.client;
+		if (newMessage.content.toLowerCase().startsWith('!')) {
+			const check = newMessage.content.substring(1, 4);
+			if (check.toLowerCase().startsWith('adm')) {
+				return;
+			}
+		}
 		if (newMessage.author.bot === true) return;
-		if (!await client.ticket.has(newMessage.channelId)) return;
-		const num = await client.ticket.get(newMessage.channelId);
-		const table = await client.db.table(`tt_${num}`);
+		if (!await lib.ticket.has(newMessage.channelId)) return;
+		const num = await lib.ticket.get(newMessage.channelId);
+		const table = await lib.db.table(`tt_${num}`);
 		const hasMessage = await table.has(newMessage.id);
 		switch (hasMessage) {
 		case true:
@@ -20,17 +27,12 @@ module.exports = {
 	},
 };
 
-async function handleHasMessage(client:any, message:any, table:Table) {
-	if (message.content.toLowerCase().startsWith('!')) {
-		const check = message.content.substring(1, 4);
-		if (check.toLowerCase().startsWith('adm')) {
-			return;
-		}
-	}
+async function handleHasMessage(client: Client, message: Message, table:QuickDB) {
 	const dataMessage = await table.get(message.id);
 	const arr = [];
 	for (const obj of dataMessage.recive) {
-		const channel = await client.channels.fetch(obj.channelId);
+		const resolvingChannel = await client.channels.fetch(obj.channelId);
+		const channel = resolvingChannel as DMChannel | TextChannel
 		const msg = await channel.messages.fetch(obj.messageId);
 		const embed = await createEmbedToSend(client, message);
 		try {
@@ -45,6 +47,11 @@ async function handleHasMessage(client:any, message:any, table:Table) {
 	if (arr.length != 0) {
 		const string = arr.join('\n');
 		message.reply({ content: 'Ni bilo mogoče spremeniti sporočila v naslednjih kanalih:\n' + string });
+	} else {
+		const embed = await createEditedEmbed(client, message);
+		const passedChannel = await client.channels.fetch(message.channelId)
+		const channel = passedChannel as TextChannel | DMChannel;
+		channel.send({ embeds: [embed] });
 	}
 }
 
@@ -52,11 +59,11 @@ async function handleNoMessage(message:Message) {
 	await message.reply({ content: 'Ni mogoče poslati spremenjenega sporočila' });
 }
 
-async function createEmbedToSend(client:any, message:any) {
+async function createEmbedToSend(client: Client, message: Message): Promise<EmbedBuilder> {
 	const user = await client.users.fetch(message.author.id);
 	const reciveChannelEmbed = new EmbedBuilder()
 		.setAuthor({ name: user.username, iconURL: user.displayAvatarURL() })
-		.setColor(await client.db.get('color.recive'))
+		.setColor(await lib.db.get('color.recive'))
 		.setTitle('Novo sporočilo')
 		.setTimestamp()
 		.setFooter({ text: `${'Ekipa BCRP'}`, iconURL: 'https://cdn.discordapp.com/attachments/1012850899980394557/1138546219640176851/097e89ede70464edaf570046b6b3f7b8.png' });
@@ -71,6 +78,17 @@ async function createEmbedToSend(client:any, message:any) {
 			num++;
 		});
 	}
+
+	return reciveChannelEmbed;
+}
+
+async function createEditedEmbed(client: Client, message: Message): Promise<EmbedBuilder> {
+	const reciveChannelEmbed = new EmbedBuilder()
+		.setColor(await lib.db.get('color.info'))
+		.setTitle('Uspešno urejeno sporočilo')
+		.setTimestamp()
+		.setFooter({ text: (message.id).toString(), iconURL: 'https://cdn.discordapp.com/attachments/1012850899980394557/1138546219640176851/097e89ede70464edaf570046b6b3f7b8.png' });
+
 
 	return reciveChannelEmbed;
 }
