@@ -1,19 +1,18 @@
-import { TextChannel } from "discord.js";
+import { TextChannel, Webhook } from "discord.js";
+import lib from '../bridge/bridge';
 
-export default webhook;
-
-async function webhook(channel:TextChannel) {
+export async function webhook(channel: TextChannel): Promise<Webhook | undefined> {
 	try {
 		let webhooks = await channel.fetchWebhooks();
 		if (webhooks.size === 0) {
-			await channel.createWebhook({ name: (channel.client as any).locales.utils.webhookjs.name, avatar: (channel.client as any).locales.utils.webhookjs.avatarURL });
+			await channel.createWebhook({ name: lib.locales.utils.webhookjs.name, avatar: lib.locales.utils.webhookjs.avatarURL });
 			webhooks = await channel.fetchWebhooks();
 		}
 		const wbh = webhooks.find((webhookurl) => webhookurl.token);
-		return (wbh);
+		return wbh;
 	}
 	catch (error) {
 		console.error(error);
-		return ('ERROR');
+		return undefined;
 	}
 }
